Fix diff_drive_skid test
A recent update to DART changed the way the slip parameter is
implemented. Part of this change removed the small amount of slip each
surface had by default. This in turn broke this test because the vehicle
would not steer as some amount of slip is needed in the lateral
direction. The solution is to lower the coefficient of friction in the
lateral direction so the friction in that direction can be overcome by
the steering force.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>