Port: Automatically load a subset of world plugins (#281) (#537)
Feature to allow loading of a default set of system plugins from a file. This behavior will trigger when a world sdf file is loaded with no plugins defined.  In this case, the simulator will load the plugins from a series of locations including environment variable, the users home folder, and finally in the installation directory.

This should allow users to not have to specify the same set of plugins in every world sdf file.

Signed-off-by: Michael Carroll <michael@openrobotics.org>
Co-authored-by: Louise Poubel <louise@openrobotics.org>