Use a custom data structure to manage entity feature maps
This ensures that when a model is removed, all the related physics
entities are removed as well causing the physics engine to clean up the
memory used by the model and its children.

This also fixes a segfault that occurs when ign-gazebo is terminated
after a model is unloaded by the level manager.  The segfault seems to
be due to a bug in DART, but the bug is exposed due to the way children
physics entity pointers outlived their parent models because they
weren't removed from their entity map when the parent model was unloaded.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>