Mecanum wheels demo (#683)
This modifies the friction parameters of the wheels in the
skid-steer example vehicle to behave like Mecanum wheels,
with anisotropic friction and friction directions that
are constant in the chassis frame. To specify these
friction directions, the //fdir1/@ignition:expressed_in
attribute is used. The friction values are specified as 1.0
in the direction corresponding to the Mecanum wheel roller
axes, and 0.0 in the perpendicular direction, which
corresponds to the rolling direction.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>