Adds velocity control to JointPositionController. (#1003)
There are times when we don't nessecarily want to tune a PID for each joint. Often we
may assume that the physical robot already has a well tuned controller. This
was true in the case of the mbari LRAUV.
Hence, there should be a way to set position without the use of a PID
controller.  This commit adds a
`use_velocity_commands` option which when set to true bypasses the force
based PID giving a very predictable
position. The `use_velocity_command` mode has a `cmd_max` which in turn
can be used to set a maximum velocity which the joint can move at.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

Co-authored-by: Louise Poubel <louise@openrobotics.org>