Prevent creation of spurious <plugin> elements when saving worlds (#1192)
Spurious <plugin> elements were created when saving worlds because sdf::Element::GetElement was used to check if a <plugin> child element exists. However this function always creates a new element if the element does not exist instead of returning a nullptr. The solution is to use sdf::Element::FindElement instead.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>