Logs a warning if a mode is not clearly sepecified. (#1307)
I had originally started this branch to fix #1305. Turns out that #1305 was not actually a problem, still I decided to take the opportunity to imporve the logging so that if a user fails to specify `<uniform_fluid_density>` or `<graded_density>` then  the user will be warned. Furthermore, I add a test for graded density mode that shows that #1305 is not an issue.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

Co-authored-by: Carlos Agüero <caguero@openrobotics.org>