Extend Multicoptor Control system to include nested model inertial params (#1450)
Updated the multicoptor velocity control system's inertial computation to include nested models. With this change, the controller will now be able to generate more thrusts when a higher motor_constant is specified. Before this change, the drone will not be able to take-off. This is useful when adding payloads (in the form of nested models) with heavy mass to the drone platform

Signed-off-by: Ian Chen <ichen@osrfoundation.org>