Fix macOS test segfaults (#1728)
The segfaults happen when components are registered by
plugins and component creation is attempted after the
plugin that registered the component has been unloaded
and it's component storage object has become invalid.
Including this header insures that all components are
registered by the core library ahead of any plugin.

This is a temporary solution. A better solution would
be to keep track of all the plugins that have registered
a component. When a plugin gets unloaded, we would
unregister that component in the destructor (right before
the plugin is unloaded), and check our list if there are
other plugins that have registered the component. If so,
we would register the component again, this time using the
component storage object from a currently loaded plugin.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>