Addresses flakiness in `Hydrodynamics.VelocityTestInOil`.
This PR adjusts the hydrodynamics tests so they are abit less flaky. Currently, it seems like the test checks if one item falls faster than another.  However due to the inherent parrallelism, it seems like on different computers the balls start falling at different times. This means the 4 time steps we wait for may not be enough for a difference in velocity to appear. I suspect we no longer need to disable this test on windows.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>