Allow SDF model to be constructed in a single shot (#1560)
This PR tweaks the behavior of the physics system so that when constructing a model it will first check if a ModelSdf component is available for the entity. If so, it uses that component to fill in all the model information, overriding the fields that are given by the other entity components (e.g. name and pose).


Signed-off-by: Michael X. Grey <grey@openrobotics.org>
Signed-off-by: Michael Carroll <michael@openrobotics.org>
Signed-off-by: ahcorde <ahcorde@gmail.com>
Co-authored-by: Michael Carroll <michael@openrobotics.org>
Co-authored-by: ahcorde <ahcorde@gmail.com>