Adds support for hydrodynamic cross terms (#1784)
* Adds support for hydrodynamic cross terms

The current plugin implementation only supports diagonal parameters. This PR adds support for cross terms in the added mass and the damping matrix.

Nothing. The plugin should continue to work as expected. Hence the test suite is not changed.

For reference read: Fossen T. "Guidance and Control of Ocean Vehicles".

Fossens equations rely on so called "stability derivatives" in the damping and the added mass terms. These are given by parameters such as `<xUU> `. What this means is `xUU` is the Quadratic Damping term (hence the repeated Us along the X axis for a velocity vector along the X axis (U, V, W are used by SNAME (Society of Naval Architects and Marine Engineers) to represent linear velocity along X, Y and Z axis repspectively). Now, in the case of many ocean going vehicles a push along the U (X-direction velocity) may result in translations and rotations in other axis (for instance my boat might yaw to the right). This may seem counter intuitive but fluid-solid interaction is often complex an unintuitive. In previous versions of the plugin, we only supported the parameters along the diagonal of the damping and added mass matrix. While this is often more than sufficient for simple simulations, there are cases where we may like to simulate cross terms. The adding of cross terms now opens up a lot more parameters. For instance, now we will have terms like `xQR` (quadratic damping in x axis with respect to angular velocity in roll and yaw axis).

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>