Fixes buoyancy flakiness when spawning entities (#1808)
* Fixes buoyancy flakiness when spawning entities

For reference see https://github.com/osrf/lrauv/pull/272#issuecomment-1323426611

Essentially, if we spawn an entity using the `UserCommand` system, it will create the entity during the `PreUpdate` phase. However, the `Buoyancy` system only checks for new entities during the `PreUpdate` phase (It does this to precompute volume so it does not need to be computed on every run). This means the buoyancy system may or may not catch the newly created entities.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>