Fix enviroment system loading mechanism
Currently, there is an issue with the way the Environment loader plugin loads data. In particular it directly writes to the ECM. While this makes sense intuitively, it does not work in practice as the GUI runs on a client process while systems that use it run on the server. This PR fixes this issue by introducing a topic through which the GUI may load Environment Data on the server.

TODO:
- [ ] This commit breaks the visuallization plugin. Fix it so the visuallization plugin runs the data sampling off the server plugin.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>