Ackermann steering with steering angle.
Option to use <steering_only> to enable a mode that does not drive the
wheels. It instead only controls the central steering angle phi in radians
on a /model/{name_of_model}/steer_angle gz::msgs::Double.
This is usefull for simulating the control of a real vehicle without
direct cmd_vel capabilities.

Signed-off-by: Benjamin Perseghetti <bperseghetti@rudislabs.com>