Document behaviour changes introduced #1784
The PR in #1784 claimed not to introduce behaviour changes,
but actually ended up introducing behaviour changes. Specifically,
it would break a lot of marine simulation code that runs on our tools.
The issue in question is a matter of naming parameters. In the past
`<xUU>` used to refer to the quadratic drag term in the diagonal axis.
This term would be multiplied by the absolute velocity. After #1784 this term
was multiplied by the velocity (no absolute). An equivalent term of `<xUabsU>`
This behaviour change likely breaks all previous maritime simulations that rely on our hydrodynamics plugin.

There are several options:
1. Revert #1784
2. Make `<xUU>` mirror `<xUabsU>`
3. Document the change and mark it as breaking.

This PR goes with option 3 as there are already code bases using `<xUabsU>` nota
tion instead of `<xUU>`. We also add a warning if someone does use the `<xUU>`
term that points to this PR for details.

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>