Use a queue to track component registration from mulitiple sources (#1836)
When a plugin gets unloaded, the component types it registered will have invalid component descriptors. Attempting to create new components of those types will result in a segfault. This PR is an attempt to fix that by keeping track of all component registrations and removing invalid ones when plugins get unloaded.

---------

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>