Speed up Resource Spawner load time by fetching model list asynchronously (#1962)
The main reason that the Resource Spawner took a long time to load is that it tried to fetch the list of all available models on Fuel instead of just the selected owner. And it did that while blocking the Qt thread, so the user was unable to interact with the GUI while the model list was being fetched. The approach I've taken is to only fetch the list of models for the default owner ("openrobotics") and allow users to add/remove any owner they want. The fetching is also done in a separate thread so as to not block the GUI.
---------

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>