Support world joints (joints inside <world> tags) (#1949)
As of gazebosim/sdformat#1117, SDFormat supports joints under world (//world/joint in xpath). This PR adds support for it in gz-sim.

Note that some systems that operate on joints, such as the JointPositionController, require that the system is instantiated inside a model, and therefore, will not work with world joints. It's not clear whether this is a real problem since world joints are most likely to be fixed joints used to attach two models together.


Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>