Adds a warning if the `Server` method of a `TestFixture` is called before `Finalize
This commit adds a warning if someone were to call `Server()` before
`Finalize`. The rationale for this is because I made the mistake of not
calling `Finalize` ina program I was writing. It took me some time to
track down the root cause of the problem. I figure if we warn users it
should be immediately obvious.

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>