Apply Force and Torque GUI plugin (#2014)
Adds a new ApplyForceTorque GUI plugin, essentially implementing the Apply Force and Torque interface that we have in Gazebo Classic.

This allows users to apply a specified force and torque to a chosen link in a model upon pressing a button. The implementation takes advantage of the existing ApplyLinkWrench system, sending the wrenches to the /world/<world_name>/wrench topic. The ApplyLinkWrench system is automatically loaded once this plugin is loaded.

---------

Signed-off-by: Henrique-BO <henrique.barrosoliveira@usp.br>