Include contact force, normal, and depth in contact message (#2050)
Previously the data contained by the ContactSensorPlugin did not include ExtraContactData. Now the message is also populated with this data. Now the EntityContactMap contains a deque of pairs of ContactPoint and ExtraContactData. That way the messages can be populated with Normals, Forces, and Depth. The force on body 2 is equal and opposite to the force on body 1.

---------

Signed-off-by: Anton Bredenbeck <a.bredenbeck@tudelft.nl>
Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
Co-authored-by: Addisu Z. Taddese <addisu@openrobotics.org>