Add force offset support to ApplyLinkWrench system and to Link API (#2026)
Adds support for specifying the application point of a force in the ApplyLinkWrench system. This is done through an offset expressed in the link frame (thus, relative to the link origin).

This uses the force_offset field that was already present in the Wrench message, but wasn't being used in the system. Also, an overloaded method AddWorldWrench was added to the Link interface.

---------

Signed-off-by: Henrique-BO <henrique.barrosoliveira@usp.br>