Force offset and vector magnitude support in ApplyForceTorque (#2056)
Adds support for specifying the offset of the force application point through the GUI in the ApplyForceTorque plugin. This offset is also reflected in the visualization of the force vector.

Also adds fields for the magnitude of the force and torque vectors on the interface. Changing the components of a vector automatically updates the value of its magnitude, while changing the vector's magnitude automatically scales its components, maintaining its direction.

---------

Signed-off-by: Henrique-BO <henrique.barrosoliveira@usp.br>