Add support for writing systems in Python (#2081)
Creates a C++ system that is able to load a python module and pass along the System interface calls, such as Configure, PreUpdate, etc. The convention is pretty simple. The python module exposes a get_system function which returns an instance of a class that implements the desired System interface functions.

---------

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>