Porting Advanced Lift Drag Plugin to Gazebo (#2185)
# Summary
This is a plugin that ports the behaviour of the advanced lift drag plugin that was present in Gazebo Classic to Gazebo. The physics implementation have not changed, but the plugin has been adapted to work with the entity component system. Primary modeling differences in the advanced_lift_drag plugin from the original liftdrag_plugin include:

- quadratic formulation for drag 
- side force
- flat-plate post-stall model
- aerodynamic moments about all three axes
- body rate stability derivatives
- actuator control derivatives

The objective is to provide a more accurate model of a wing than what is provided in the basic lift drag plugin.

Signed-off-by: frederik <frederik@auterion.com>
Signed-off-by: Frederik Markus <80588263+frede791@users.noreply.github.com>
Signed-off-by: Arjo Chakravarty <arjo129@gmail.com>
Co-authored-by: frederik <frederik@auterion.com>
Co-authored-by: Arjo Chakravarty <arjo129@gmail.com>