Default CMA in LiftDrag pluginto zero
The changes in #2189 have caused a regression for our upstream users.
This fix disables moment calculations by defaulting Cma to zero. It
should help mitigate some of the regression. There is also an [ongoing
discussion](https://github.com/gazebosim/gz-sim/pull/2189#issuecomment-1866589946) as to whether the cos term should be cos^2 or cos.

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>