Fix namespace and class links in documentation references that use namespace `gz`
The `gz` namespace is not recognized by doxygen enough for it to
autogenerate links to inner namespaces and classes. For example, `\ref
gz::sim::systems` or simply `gz::sim::systems` don't link to the
namespace documentation. You'd have to use `ignition::gazebo::systems` instead. I was hoping there
was a more clever solution that would allow all `gz::` refs to work
automatically, but I was not able to find one, so this is a compromise
to fix the landing page for ign-gazebo6 (https://gazebosim.org/api/sim/6)

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>