LTA Dynamics System (#2241)
This adds the dynamics to simulate a lighter-than-air vehicle such as blimps/airships. The modelling is  based on [1] and [2]. Previously we added an airship-dynamics-plugin to Gazebo-Classic for [PX4 Gazebo SITL](https://github.com/PX4/PX4-SITL_gazebo-classic/pull/490) , but we suffered simulation instability similar to that of the underwater-vehicles since Gazebo-Classic could not simulate the Added Mass dynamics.

This PR brings that airship-dynamics-plugin to Gazebo which supports the Added Mass dynamics.

Furthermore, I have extended, the Link class to provide access to the AddedMassMatrix, to allow the lighter-than-air system to easily access this matrix. 

The code is somewhat based of the Hydrodynamics system. 

I have an STL file which I can contribute, which is an envelope from [Wind Reiter](https://www.windreiter.com/shop/sb-324-300/), and the coefficients in the integrations test is for this envelope model.

### Citations
[1] Li, Y., & Nahon, M. (2007). Modeling and simulation of airship dynamics.  Journal of Guidance, Control, and Dynamics, 30(6), 1691–1700.
[2] Li, Y., Nahon, M., & Sharf, I. (2011). Airship dynamics modeling: A literature review. Progress in Aerospace Sciences, 47(3), 217–239.


Signed-off-by: henrykotze <henry@flycloudline.com>