Add SystemPriorityConstants.hh header file
Define a constant for the Physics system priority
so that it runs before systems with default priority
and use it in the Physics system.

Define a constant for a sensor system priority that
runs after Physics but before systems with default
priority and use it in the ForceTorque system.

Define a constant suggesting a priority for systems
to be executed before Physics.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>