Provides step size via CLI interface and `ServerConfig`
This is probably not the best time to be making this PR as we are in the
middle of the release. However for another project, I'd like to be able
to set the Step Size via CLI. This is particularly useful for when we
are doing integration testing of that project. This PR provides two
command line flags that allow you to set the physics update rate and
step size via cli. To try it out run:

```
gz sim shapes.sdf --physics-rtf 5.0 --physics-step-size 0.01
```

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>