A hack to greatly improve load times
I was investigating
https://github.com/gazebosim/gazebo_test_cases/issues/1576 , in my
investigation it came to my notice that `sdf::Element` takes forever to
destroy (We should open a ticket somewhere about this). If we are
skipping serialization we might as well not create and destroy an
SDF Element. This hack greatly speeds up the load time for gazebo.

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>