Adds support for `ISystemReset` in test fixture (#2647)
* Adds support for Reset in test fixture

This PR adds support for the Reset API to the test fixture. As `TestFixture`
is one of the main ways one can get access to the ECM in python
when trying to write some scripts for Deep Reinforcement Learning I
realized that without `Reset` supported in the `TestFixture` API, end
users would have a very hard time using our python APIs (which are
actually quite nice). For reference I'm hacking a demo template here:

https://github.com/gazebosim/gz-sim/pull/2667
---------

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>