Attempts to print a stacktrace on segfault in our integration tests (#2776)
* Attempts to print a stacktrace on segfault in our integration tests

Our integration tests often segfault in CI leaving us clueless as to
whats happening. This commit is an attempt to vendor backward-cpp and
at the same time allows us to get a stacktrace of our code. The current
status of this PR is draft.

Other options considered:

We vendor backward-cpp in `gz-launch` and in `gz-utils`. Sadly the
ubuntu package is not very useful as no cmake targets are provided. I
could try to vendor it in gz-cmake or write a besopoke `FindBackward`
package. However, our goal is to use it in tests. This is a very
specific use case and I feel maybe the simplest option is to vendor it
here and link it against the tests. That way we do not introduce more
dependencies via 3rd party package managers and the user has no need to
have anything else installed.

---------

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>
Co-authored-by: Jose Luis Rivero <jrivero@osrfoundation.org>
Co-authored-by: Carlos Agüero <caguero@openrobotics.org>
Co-authored-by: Jose Luis Rivero <jrivero@openrobotics.com>
(cherry picked from commit 7608ed9dc479344a7815e704b59bbd9694152ed4)

# Conflicts:
#	.github/workflows/ci.yml