Link::AddForceInInertialFrame APIs added (#2816)
This adds a pair of overloaded APIs for applying forces
to a Link specified in coordinates of its Inertial frame,
one of which applies the force directly at the center of
mass and the other applies the force at a specified
offset from the center of mass. For consistency, all
vector parameters to these APIs are relative to the Inertial
frame and expressed in coordinates of the Inertial frame.

Signed-off-by: Vedant87 <Vedantsr3@gmail.com>
Signed-off-by: Vedant Randive <vedantsr3@gmail.com>
Co-authored-by: Steve Peters <scpeters@openrobotics.org>