Physics System populates link bounding boxes (#2821)
Computes the link bounding boxes for all link entities that have components::AxisAlignedBox present. The computation follows the same logic as the models bbox, but falls back to the SDF calculation introduced in #2787 in case the feature is not present for the selected physics plugin.

---------

Signed-off-by: Gabriel Pacheco <gabriel.fpacheco@gmail.com>
Co-authored-by: Ian Chen <ichen@openrobotics.org>
(cherry picked from commit 716de42bcd18901940df29044a73ebaf2421c720)