Update our usage of workerpools (#2995)
In `SimulationRunner`, we initialize a worker pool but we never actually
use it so it's pure overhead. In `ServerPrivate` we only use the
worker pool if there are multiple simulation runners, so we can optimize
for the most common use case of one runner.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
(cherry picked from commit 5ce62a8826b930078fae940288ac9677aae9ec7b)