Refactor Transport service handlers `UserCommands` (#2996)
* Refactor Transport service handlers

This removes a lot of repetitive code where a service handler simply
copies the message, creates and adds the associated command to a command
queue.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Also refactor code for advertising services

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Add back missing function

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Update our usage of workerpools (#2995)

In `SimulationRunner`, we initialize a worker pool but we never actually
use it so it's pure overhead. In `ServerPrivate` we only use the
worker pool if there are multiple simulation runners, so we can optimize
for the most common use case of one runner.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Fix linter issues

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Fix typo

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Add missing constructor

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

---------

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
(cherry picked from commit bf68eaead4935558ae90fb0c5c7e9544420c652d)