Expose APIs for clearing the ECMs internal state tracking addition and removal (#3002)
This is necessary for use with the `SetState` API. This API is commonly
used to create a local version of the servers ECM that is periodically
synchronized. However, without exposing this function into our public API
additions and removals of entities or components are not reflected in the
local ECM.

Currently, we use `friend` relationships for the `GuiRunner` class to
allow it to run these functions, but this is not feasible for external
code that wants to use `SetState`.

https://github.com/gazebosim/gz-sim/blob/5ce62a8826b930078fae940288ac9677aae9ec7b/include/gz/sim/EntityComponentManager.hh#L834,
https://github.com/gazebosim/gz-sim/blob/5ce62a8826b930078fae940288ac9677aae9ec7b/src/gui/GuiRunner.cc#L316-L318

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
(cherry picked from commit b2f3c82288760ea9a7819a3dc404c05eba47af9a)