Remove server+client flags from independent standalone execs (#2990)
The PR removes from the standalone executables the parameters (--render-engine and --render-engine-api-backend) that affects the server and client. This parameter makes sense to be used when the gz sim combined utility is being run to affect both part of the application. The standalone executables ship the customized versions of these  --render-engine-server and --render-engine-gui.

Signed-off-by: Jose Luis Rivero <jrivero@honurobotics.com>
Co-authored-by: Addisu Z. Taddese <addisu@openrobotics.org>